requirements.txt
python-dotenv
google-generativeai
flask
faiss-cpu 
numpy
sentence-transformers
requests